+++ mycodo/utils/functions.py
@@ -61,7 +61,7 @@ def parse_function_information(exclude_custom=False):
                 continue
 
             full_path = "{}/{}".format(real_path, each_file)
-            function_custom = load_module_from_file(full_path, 'functions')
+            function_custom, status = load_module_from_file(full_path, 'functions')
 
             if not function_custom or not hasattr(function_custom, 'FUNCTION_INFORMATION'):
                 continue