+++ mycodo/utils/widgets.py
@@ -70,7 +70,7 @@ def parse_widget_information(exclude_custom=False):
                 continue
 
             full_path = "{}/{}".format(real_path, each_file)
-            widget_custom = load_module_from_file(full_path, 'widgets')
+            widget_custom, status = load_module_from_file(full_path, 'widgets')
 
             if not widget_custom or not hasattr(widget_custom, 'WIDGET_INFORMATION'):
                 continue